feat: define road borders, set camera view on top of the car
Initially we built a moving, car. So the animation was about a rectangular
able to move around. Despite this is convenient for the physics it is not for
the final representation.
To solve that we updated animation to translate canvas relatively to the moving
car coordinates (y specifically). Now we have this notion of the moving road instead
of the a moving car...