feat: add car polygon
Until now we used to consider car to be fixed in terms of
rotation and shape rectangular. With the current update we do
introduce the notion of polygon.

Given the center of the car we draw a polygon around it.
Instead of just translating the canvas to any polygon position/angle
update we just plot the new polygon mesh around.

In this notion we have sufficient information to be able to detect
collisions or to model any car damage in the case of some crash.